LayoutTests/svg/dom/SVGSVGElement-currentTime.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="200" height="200"></svg>
<script>
description("Calling svgElement.setCurrentTime() with a non-finite value should throw.");

self.jsTestIsAsync = true;

if (window.testRunner) {
     testRunner.dumpAsText();
     testRunner.waitUntilDone();
}

function runTest()
{
    svgElement = document.getElementById("svg");
    shouldThrow("svgElement.setCurrentTime(NaN)");
    shouldThrow("svgElement.setCurrentTime(Infinity)");
    finishJSTest();
}
window.onload = runTest;
</script>
</body>
</html>
