LayoutTests/svg/dom/SVGSVGElement-currentTime-expected.txt

Calling svgElement.setCurrentTime() with a non-finite value should throw.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgElement.setCurrentTime(NaN) threw exception TypeError: The provided value is non-finite.
PASS svgElement.setCurrentTime(Infinity) threw exception TypeError: The provided value is non-finite.
PASS successfullyParsed is true

TEST COMPLETE
